requirements.txt
google-genai
python-dotenv
webuiapi
Pillow
